requirements.txt
undetected_chromedriver
pyyaml
colorama